+++ src/hcj.js
@@ -1703,6 +1703,15 @@
       var mhS = stream.create();
       var spanStreams = [];
       var pushingDimensions = false;
+      var innerEl;
+      if (config.underline || config.overline || config.lineThrough) {
+        innerEl = document.createElement('div');
+        innerEl.style.height = '100%';
+        el.appendChild(innerEl);
+      }
+      else {
+        innerEl = el;
+      }
       var pushDimensions = function () {
         if (pushingDimensions) {
           return;
@@ -1839,11 +1848,11 @@
             a.classList.add('no-style');
           }
           a.href = c.linkTo;
-          el.appendChild(a);
+          innerEl.appendChild(a);
           a.appendChild(span);
         }
         else {
-          el.appendChild(span);
+          innerEl.appendChild(span);
         }
       });
 
@@ -1871,6 +1880,14 @@
           }
         }
       });
+
+      if (config.underline || config.overline || config.lineThrough) {
+        var underlineStyle = config.underline;
+        if (underlineStyle === true) {
+          underlineStyle = 'solid';
+        }
+        innerEl.style.textDecoration = underlineStyle + ' underline';
+      }
 
       if (config.spanCSS) {
         config.spanCSS.map(function (css) {